refactor: simplify Avatar component by utilizing custom useAvatar hook
